Few signedness warning fixed.
git-svn-id: https://nicola-bonelli-repo.googlecode.com/svn@935 0daf1152-ea34-11dd-b00f-9fabb4b2ff13

+++ trunk/codes.cpp11/socket-test.cpp
@@ -25,7 +25,7 @@ void client(std::string host, unsigned short port)
 
     const char hello_world[] = "hello world!";
 
-    auto len = htonl(static_cast<int>(sizeof(hello_world)));  // this include '\0'
+    auto len = htonl(sizeof(hello_world));  // this include '\0'
 
     sock.send_atomic(more::const_buffer(reinterpret_cast<const char *>(&len), sizeof(len)), 0);
     sock.send_atomic(more::const_buffer(hello_world, sizeof(hello_world)), 0);
